chore: 🔧 Create a unified store & limit scheduling overdrag (#317)
* chore: 🔧 created new store for unified API, combined all states/actions from previous 6 stores

* chore: 🔧 replace old stores in availability.tsx

* chore: 🔧 replace old stores in group-availability.tsx

* chore: 🔧 replaced sub-stores with unified store, remove subcomponent state call in async handler

* chore: 🔧 replace old stores with unified store in group-responses

* chore: 🔧 remove & replace sub stores in personal availability

* chore: 🔧 replace sub stores in availability block cell

* chore: 🔧 replace old stores in availability blocks

* chore: 🔧 replace old stores in hooks

* chore: 🔧 delete all sub stores (now combined into useAvailabilityStore)

* chore: 🔧 removed redundant userAvailability prop

* chore: avoid re-rendering best times for unrelated store updates


* chore: 🔧 remove unnecessary store usage and restrict prevPage to only update state when current page>0

* chore: use selector to read from unified store to avoid unnecessary rerender


* chore: 🔧 readded replaceEntireSelection function after merge conflicts

* fix: 🐛 replace previous scheduled block on new drag

* fix: 🐛 adding constraints so overdragging in scheduling mode can't schedule over multiple days

* chore: 🔧 rm unused props

* chore: 🔧 rm usused selectors

---------
