+++ src/components/availability/availability.tsx
@@ -40,25 +40,24 @@ import {
 	isAnchorDateMeeting,
 } from "@/lib/types/chrono";
 import { ZotDate } from "@/lib/zotdate";
-import { useAvailabilityPaginationStore } from "@/store/useAvailabilityPaginationStore";
-import { useAvailabilityViewStore } from "@/store/useAvailabilityViewStore";
-import { useBlockSelectionStore } from "@/store/useBlockSelectionStore";
-import { useGroupSelectionStore } from "@/store/useGroupSelectionStore";
-import { useScheduleSelectionStore } from "@/store/useScheduleSelectionStore";
+import { useAvailabilityStore } from "@/store/useAvailabilityStore";
 
 // Helper function to derive initial availability data
 const deriveInitialAvailability = ({
 	meetingDates,
-	userAvailability,
+	userId,
 	allAvailabilties,
 	availabilityTimeBlocks,
 }: {
 	timezone: string;
 	meetingDates: string[];
-	userAvailability: MemberMeetingAvailability | null;
+	userId: string | null;
 	allAvailabilties: MemberMeetingAvailability[];
 	availabilityTimeBlocks: number[];
 }) => {
+	const userAvailability =
+		allAvailabilties.find((a) => a.memberId === userId) ?? null;
+
 	const availabilitiesByDate = new Map<string, string[]>();
 	if (userAvailability?.meetingAvailabilities) {
 		userAvailability.meetingAvailabilities.forEach((timeStr) => {
@@ -128,13 +127,11 @@ const deriveInitialAvailability = ({
 
 export function Availability({
 	meetingData,
-	userAvailability,
 	allAvailabilities,
 	user,
 	scheduledBlocks,
 }: {
 	meetingData: SelectMeeting;
-	userAvailability: MemberMeetingAvailability | null;
 	allAvailabilities: MemberMeetingAvailability[];
 	user: UserProfile | null;
 	scheduledBlocks: SelectScheduledMeeting[];
@@ -142,27 +139,25 @@ export function Availability({
 	const {
 		availabilityView,
 		setAvailabilityView,
-		setHasAvailability,
 		hasAvailability,
-	} = useAvailabilityViewStore(
+		setHasAvailability,
+	} = useAvailabilityStore(
 		useShallow((state) => ({
 			availabilityView: state.availabilityView,
 			setAvailabilityView: state.setAvailabilityView,
-			setHasAvailability: state.setHasAvailability,
 			hasAvailability: state.hasAvailability,
+			setHasAvailability: state.setHasAvailability,
 		})),
 	);
 
-	const selectionIsLocked = useGroupSelectionStore(
+	const selectionIsLocked = useAvailabilityStore(
 		(state) => state.selectionIsLocked,
 	);
-	const resetSelection = useGroupSelectionStore(
-		(state) => state.resetSelection,
-	);
-	const setIsMobileDrawerOpen = useGroupSelectionStore(
+	const resetSelection = useAvailabilityStore((state) => state.resetSelection);
+	const setIsMobileDrawerOpen = useAvailabilityStore(
 		(state) => state.setIsMobileDrawerOpen,
 	);
-	const toggleHoverGrid = useGroupSelectionStore(
+	const toggleHoverGrid = useAvailabilityStore(
 		(state) => state.toggleHoverGrid,
 	);
 
@@ -181,7 +176,7 @@ export function Availability({
 	]);
 
 	const { currentPage, itemsPerPage, isFirstPage, nextPage, prevPage } =
-		useAvailabilityPaginationStore(
+		useAvailabilityStore(
 			useShallow((state) => ({
 				currentPage: state.currentPage,
 				itemsPerPage: state.itemsPerPage,
@@ -282,7 +277,7 @@ export function Availability({
 		deriveInitialAvailability({
 			timezone: userTimezone,
 			meetingDates: meetingData.dates,
-			userAvailability,
+			userId: user?.memberId ?? null,
 			allAvailabilties: allAvailabilities,
 			availabilityTimeBlocks,
 		}),
@@ -499,7 +494,7 @@ export function Availability({
 		}
 
 		// add DB timestamps to the state
-		useScheduleSelectionStore.getState().hydrateScheduledTimes(timestamps);
+		useAvailabilityStore.getState().hydrateScheduledTimes(timestamps);
 	}, [scheduledBlocks]);
 
 	// Drag selection for group and schedule views
@@ -509,7 +504,7 @@ export function Availability({
 		setStartBlockSelection,
 		setEndBlockSelection,
 		setSelectionState,
-	} = useBlockSelectionStore(
+	} = useAvailabilityStore(
 		useShallow((state) => ({
 			startBlockSelection: state.startBlockSelection,
 			endBlockSelection: state.endBlockSelection,
@@ -519,14 +514,11 @@ export function Availability({
 		})),
 	);
 
-	const { togglePendingTime, addPendingTimeRange, isScheduled } =
-		useScheduleSelectionStore(
-			useShallow((state) => ({
-				togglePendingTime: state.togglePendingTime,
-				addPendingTimeRange: state.addPendingTimeRange,
-				isScheduled: state.isScheduled,
-			})),
-		);
+	const { replaceEntireSelection } = useAvailabilityStore(
+		useShallow((state) => ({
+			replaceEntireSelection: state.replaceEntireSelection,
+		})),
+	);
 
 	useEffect(() => {
 		if (startBlockSelection && endBlockSelection) {
@@ -593,7 +585,7 @@ export function Availability({
 				setEndBlockSelection({ zotDateIndex, blockIndex });
 			} else if (last) {
 				const { startBlockSelection, endBlockSelection } =
-					useBlockSelectionStore.getState();
+					useAvailabilityStore.getState();
 				if (startBlockSelection && endBlockSelection) {
 					const earlierDateIndex = Math.min(
 						startBlockSelection.zotDateIndex,
@@ -613,38 +605,26 @@ export function Availability({
 					);
 
 					if (availabilityView === "schedule") {
+						// Schedule selection stays on the day the drag started (no cross-day rectangle).
+						const day = startBlockSelection.zotDateIndex;
 						const timestamps: string[] = [];
 						for (
-							let dateIndex = earlierDateIndex;
-							dateIndex <= laterDateIndex;
-							dateIndex++
+							let blockI = earlierBlockIndex;
+							blockI <= laterBlockIndex;
+							blockI++
 						) {
-							for (
-								let blockI = earlierBlockIndex;
-								blockI <= laterBlockIndex;
-								blockI++
-							) {
-								const timestamp = getTimestampFromBlockIndex(
-									blockI,
-									dateIndex,
-									fromTimeMinutes,
-									availabilityDates,
-								);
-								if (timestamp) timestamps.push(timestamp);
-							}
+							const timestamp = getTimestampFromBlockIndex(
+								blockI,
+								day,
+								fromTimeMinutes,
+								availabilityDates,
+							);
+							if (timestamp) timestamps.push(timestamp);
 						}
 
 						if (timestamps.length > 0) {
-							const firstTimestamp = timestamps[0];
-							const isFirstScheduled = isScheduled(firstTimestamp);
-
-							if (isFirstScheduled) {
-								timestamps.forEach((ts) => {
-									if (isScheduled(ts)) togglePendingTime(ts);
-								});
-							} else {
-								addPendingTimeRange(timestamps);
-							}
+							// Each drag replaces the full scheduled selection (one contiguous rectangle).
+							replaceEntireSelection(timestamps);
 						}
 					} else if (availabilityView === "personal") {
 						const startZotDate =
@@ -714,6 +694,16 @@ export function Availability({
 			} else {
 				if (blockInfo) {
 					const { zotDateIndex, blockIndex } = blockInfo;
+					if (availabilityView === "schedule") {
+						const start = useAvailabilityStore.getState().startBlockSelection;
+						if (start) {
+							setEndBlockSelection({
+								zotDateIndex: start.zotDateIndex,
+								blockIndex,
+							});
+							return;
+						}
+					}
 					setEndBlockSelection({ zotDateIndex, blockIndex });
 				}
 			}
