+++ src/components/availability/header/availability-header.tsx
@@ -27,8 +27,7 @@ import type { UserProfile } from "@/lib/auth/user";
 import { formatTimeWithHoursAndMins } from "@/lib/availability/utils";
 import { cn } from "@/lib/utils";
 import type { ZotDate } from "@/lib/zotdate";
-import { useAvailabilityViewStore } from "@/store/useAvailabilityViewStore";
-import { useScheduleSelectionStore } from "@/store/useScheduleSelectionStore";
+import { useAvailabilityStore } from "@/store/useAvailabilityStore";
 
 interface AvailabilityHeaderProps {
 	meetingData: SelectMeeting;
@@ -51,7 +50,7 @@ export function AvailabilityHeader({
 }: AvailabilityHeaderProps) {
 	const router = useRouter();
 
-	const { availabilityView, setAvailabilityView } = useAvailabilityViewStore(
+	const { availabilityView, setAvailabilityView } = useAvailabilityStore(
 		useShallow((state) => ({
 			availabilityView: state.availabilityView,
 			setAvailabilityView: state.setAvailabilityView,
@@ -83,7 +82,7 @@ export function AvailabilityHeader({
 		onSave();
 	};
 
-	const { commitPendingTimes, clearPendingTimes } = useScheduleSelectionStore(
+	const { commitPendingTimes, clearPendingTimes } = useAvailabilityStore(
 		useShallow((state) => ({
 			pendingAdds: state.pendingAdds,
 			commitPendingTimes: state.commitPendingTimes,
@@ -98,8 +97,7 @@ export function AvailabilityHeader({
 
 	const handleScheduleSave = async () => {
 		try {
-			const { pendingAdds, pendingRemovals } =
-				useScheduleSelectionStore.getState();
+			const { pendingAdds, pendingRemovals } = useAvailabilityStore.getState();
 
 			// Remove pending removals
 			for (const timestamp of pendingRemovals) {
@@ -121,7 +119,6 @@ export function AvailabilityHeader({
 					scheduledToTime,
 				});
 			}
-
 			// Add new pending times
 			for (const timestamp of pendingAdds) {
 				const date = new Date(timestamp);
