+++ src/components/availability/table/availability-blocks.tsx
@@ -8,7 +8,7 @@ import {
 } from "@/lib/availability/utils";
 import type { ProcessedCellEventSegments } from "@/lib/types/availability";
 import type { ZotDate } from "@/lib/zotdate";
-import { useAvailabilityPaginationStore } from "@/store/useAvailabilityPaginationStore";
+import { useAvailabilityStore } from "@/store/useAvailabilityStore";
 
 interface AvailabilityBlocksProps {
 	timeBlock: number;
@@ -25,7 +25,7 @@ export function AvailabilityBlocks({
 	currentPageAvailability,
 	processedCellSegments,
 }: AvailabilityBlocksProps) {
-	const { currentPage, itemsPerPage } = useAvailabilityPaginationStore(
+	const { currentPage, itemsPerPage } = useAvailabilityStore(
 		useShallow((state) => ({
 			currentPage: state.currentPage,
 			itemsPerPage: state.itemsPerPage,
